[simulation] optimisations et cleaning zone urbaine

+++ src/libraries/simulation_kernel/Entities/Agents/Perceptions/PHY_PerceptionSurfaceAgent.cpp
@@ -244,11 +244,10 @@ const std::string& PHY_PerceptionSurfaceAgent::GetSensorTypeName() const
 // Name: PHY_PerceptionSurfaceAgent::TransfertPerception
 // Created: SLG 2010-05-07
 // -----------------------------------------------------------------------------
-void PHY_PerceptionSurfaceAgent::TransfertPerception( std::map< const void*, std::pair< unsigned int, float > > urbanPerceptionMap ) const
+void PHY_PerceptionSurfaceAgent::TransfertPerception( std::map< const void*, std::pair< unsigned int, double > > urbanPerceptionMap ) const
 {
     perceptionsBuffer_.clear();
-
-    for ( std::map< const void*, std::pair< unsigned int, float > >::const_iterator it = urbanPerceptionMap.begin(); it != urbanPerceptionMap.end(); ++it )
+    for( std::map< const void*, std::pair< unsigned int, double > >::const_iterator it = urbanPerceptionMap.begin(); it != urbanPerceptionMap.end(); ++it )
         perceptionsBuffer_[ it->first ] = it->second.first;
 }
